move logic to securitylogic
