+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/security/OrderlyWebSecurityLogic.kt
@@ -7,6 +7,10 @@ import org.pac4j.core.http.adapter.HttpActionAdapter
 import org.pac4j.core.profile.CommonProfile
 import org.pac4j.core.profile.ProfileManager
 import org.pac4j.sparkjava.SparkWebContext
+import org.vaccineimpact.orderlyweb.db.OrderlyAuthorizationRepository
+import org.vaccineimpact.orderlyweb.models.Scope
+import org.vaccineimpact.orderlyweb.models.permissions.ReifiedPermission
+import org.vaccineimpact.orderlyweb.security.authorization.orderlyWebPermissions
 import org.vaccineimpact.orderlyweb.security.clients.GithubIndirectClient
 
 class OrderlyWebSecurityLogic : DefaultSecurityLogic<Any?, SparkWebContext>()
@@ -22,10 +26,29 @@ class OrderlyWebSecurityLogic : DefaultSecurityLogic<Any?, SparkWebContext>()
                          vararg parameters: Any?): Any?
     {
         val client = clientFinder.find(config?.clients, context, clients).first()
+        val manager = ProfileManager<CommonProfile>(context)
         if (client is GithubIndirectClient)
         {
             // wipe anon profile
-            ProfileManager<CommonProfile>(context).remove(true)
+            manager.remove(true)
+        }
+        else
+        {
+            val currentProfile = manager.get(true)
+
+            if (!currentProfile.isPresent || currentProfile.get().id == "anon")
+            {
+                manager.remove(true)
+                val profile = CommonProfile().apply { id = "anon" }
+
+                val permissions = OrderlyAuthorizationRepository()
+                        .getPermissionsForUser("anon")
+
+                profile.orderlyWebPermissions = permissions.plus(ReifiedPermission("reports.read",
+                        Scope.Specific("report", "minimal")))
+
+                manager.save(true, profile, false)
+            }
         }
         return super.perform(context,
                 config,
